test(server): start testing happy path
