+++ packages/twenty-server/test/integration/utils/object-metadata/generate-mock-create-object-metadata-input.ts
@@ -1,0 +1,15 @@
+import { CreateObjectInput } from 'src/engine/metadata-modules/object-metadata/dtos/create-object.input';
+
+// TODO use faker
+export const getMockCreateObjectInput = (
+  overrides?: Partial<Omit<CreateObjectInput, 'workspaceId' | 'dataSourceId'>>,
+) => ({
+  namePlural: 'listings',
+  nameSingular: 'toto',
+  labelPlural: 'Listings',
+  labelSingular: 'Listing',
+  description: 'Listing object',
+  icon: 'IconListNumbers',
+  isLabelSyncedWithName: false,
+  ...overrides,
+});
